add a delegating handler to put the encapsulate the noice
