+++ beacon_node/eth2-libp2p/src/behaviour/mod.rs
@@ -5,7 +5,7 @@ use crate::types::{GossipEncoding, GossipKind, GossipTopic};
 use crate::{error, Enr, NetworkConfig, NetworkGlobals, PubsubMessage, TopicHash};
 use discv5::Discv5Event;
 use futures::prelude::*;
-use handler::{BHInEvent, BehaviourHandler, DelegateIn};
+use handler::{BehaviourHandler, BehaviourHandlerIn, BehaviourHandlerOut, DelegateIn, DelegateOut};
 use libp2p::{
     core::{
         connection::{ConnectedPoint, ConnectionId, ListenerId},
@@ -51,6 +51,7 @@ pub struct Behaviour<TSpec: EthSpec> {
     peer_manager: PeerManager<TSpec>,
     /// The events generated by this behaviour to be consumed in the swarm poll.
     events: Vec<BehaviourEvent<TSpec>>,
+    // TODO: add events to send to the handler
     /// The current meta data of the node, so respond to pings and get metadata
     meta_data: MetaData<TSpec>,
     /// A cache of recently seen gossip messages. This is used to filter out any possible
@@ -170,16 +171,15 @@ impl<TSpec: EthSpec> NetworkBehaviour for Behaviour<TSpec> {
     ) {
         // Events comming from the handler, redirected to each behaviour
         match event {
-            EitherOutput::First(EitherOutput::First(EitherOutput::First(ev))) => {
-                self.gossipsub.inject_event(peer_id, conn_id, ev)
+            BehaviourHandlerOut::Delegate(delegate) => match delegate {
+                DelegateOut::Gossipsub(ev) => self.gossipsub.inject_event(peer_id, conn_id, ev),
+                DelegateOut::RPC(ev) => self.eth2_rpc.inject_event(peer_id, conn_id, ev),
+                DelegateOut::Identify(ev) => self.identify.inject_event(peer_id, conn_id, ev),
+                DelegateOut::Discovery(ev) => self.discovery.inject_event(peer_id, conn_id, ev),
+            },
+            BehaviourHandlerOut::Custom => {
+                // TODO: implement
             }
-            EitherOutput::First(EitherOutput::First(EitherOutput::Second(ev))) => {
-                self.eth2_rpc.inject_event(peer_id, conn_id, ev)
-            }
-            EitherOutput::First(EitherOutput::Second(ev)) => {
-                self.identify.inject_event(peer_id, conn_id, ev)
-            }
-            EitherOutput::Second(ev) => self.discovery.inject_event(peer_id, conn_id, ev),
         }
     }
 
@@ -218,9 +218,9 @@ impl<TSpec: EthSpec> NetworkBehaviour for Behaviour<TSpec> {
                                     handler,
                                     // call the closure mapping the received event to the needed one
                                     // in order to notify the handler
-                                    event: BHInEvent::Delegate($notify_handler_event_closure(
-                                        event,
-                                    )),
+                                    event: BehaviourHandlerIn::Delegate(
+                                        $notify_handler_event_closure(event),
+                                    ),
                                 });
                             }
                             NBAction::ReportObservedAddr { address } => {
